Move all tests with Tomcat server in a separate class
Start Tomcat just once for all tests.
