+++ src/test/java/reactor/netty/http/client/HttpClientTest.java
@@ -82,7 +82,6 @@ import reactor.netty.DisposableServer;
 import reactor.netty.FutureMono;
 import reactor.netty.NettyPipeline;
 import reactor.netty.SocketUtils;
-import reactor.netty.TomcatServer;
 import reactor.netty.channel.AbortedException;
 import reactor.netty.http.server.HttpServer;
 import reactor.netty.resources.ConnectionProvider;
@@ -319,251 +318,6 @@ public class HttpClientTest {
 
 		latch.await();
 		c.disposeNow();
-	}
-
-	@Test
-	public void postUpload() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		HttpClient client =
-				HttpClient.create()
-				          .tcpConfiguration(tcpClient -> tcpClient.host("localhost"))
-				          .port(tomcat.port())
-				          .wiretap(true);
-
-		Tuple2<Integer, String> res;
-		try (InputStream f = getClass().getResourceAsStream("/smallFile.txt")) {
-			res = client.post()
-			      .uri("/multipart")
-			      .sendForm((req, form) -> form.multipart(true)
-			                                   .file("test", f)
-			                                   .attr("attr1", "attr2")
-			                                   .file("test2", f))
-			      .responseSingle((r, buf) -> buf.asString().map(s -> Tuples.of(r.status().code(), s)))
-			      .block(Duration.ofSeconds(30));
-		}
-
-		assertThat(res).as("response").isNotNull();
-		assertThat(res.getT1()).as("status code").isEqualTo(200);
-		assertThat(res.getT2()).as("response body reflecting request").contains("test: 95\nattr1: 5\ntest2: 0\n");
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void simpleTest404() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		doSimpleTest404(HttpClient.create()
-		                          .baseUrl("http://localhost:" + tomcat.port()));
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void simpleTest404_1() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		ConnectionProvider pool = ConnectionProvider.fixed("http", 1);
-		HttpClient client =
-				HttpClient.create(pool)
-				          .port(tomcat.port())
-				          .tcpConfiguration(tcpClient -> tcpClient.host("localhost"))
-				          .wiretap(true);
-		doSimpleTest404(client);
-		doSimpleTest404(client);
-		pool.dispose();
-
-		tomcat.stop();
-	}
-
-	private void doSimpleTest404(HttpClient client) {
-		Integer res = client.followRedirect(true)
-				            .get()
-				            .uri("/status/404")
-				            .responseSingle((r, buf) -> Mono.just(r.status().code()))
-				            .log()
-				            .block();
-
-		assertThat(res).isNotNull();
-		if (res != 404) {
-			throw new IllegalStateException("test status failed with " + res);
-		}
-	}
-
-	@Test
-	public void disableChunkForced() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		AtomicReference<HttpHeaders> headers = new AtomicReference<>();
-		Tuple2<HttpResponseStatus, String> r =
-				HttpClient.newConnection()
-				          .tcpConfiguration(tcpClient -> tcpClient.host("localhost"))
-				          .port(tomcat.port())
-				          .headers(h -> h.set(HttpHeaderNames.TRANSFER_ENCODING, HttpHeaderValues.CHUNKED))
-				          .wiretap(true)
-				          .doAfterRequest((req, connection) -> headers.set(req.requestHeaders()))
-				          .request(HttpMethod.GET)
-				          .uri("/status/400")
-				          .send(ByteBufFlux.fromString(Flux.just("hello")))
-				          .responseSingle((res, conn) -> Mono.just(res.status())
-				                                             .zipWith(conn.asString()))
-				          .block(Duration.ofSeconds(30));
-
-		assertThat(r).isNotNull();
-
-		assertThat(r.getT1()).isEqualTo(HttpResponseStatus.BAD_REQUEST);
-		assertThat(headers.get().get("Content-Length")).isEqualTo("5");
-		assertThat(headers.get().get("Transfer-Encoding")).isNull();
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void disableChunkForced2() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		AtomicReference<HttpHeaders> headers = new AtomicReference<>();
-		Tuple2<HttpResponseStatus, String> r =
-				HttpClient.newConnection()
-				          .tcpConfiguration(tcpClient -> tcpClient.host("localhost"))
-				          .port(tomcat.port())
-				          .wiretap(true)
-				          .doAfterRequest((req, connection) -> headers.set(req.requestHeaders()))
-				          .keepAlive(false)
-				          .get()
-				          .uri("/status/404")
-				          .responseSingle((res, conn) -> Mono.just(res.status())
-				                                             .zipWith(conn.asString()))
-				          .block(Duration.ofSeconds(30));
-
-		assertThat(r).isNotNull();
-
-		assertThat(r.getT1()).isEqualTo(HttpResponseStatus.NOT_FOUND);
-		assertThat(headers.get().get("Content-Length")).isEqualTo("0");
-		assertThat(headers.get().get("Transfer-Encoding")).isNull();
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void simpleClientPooling() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		ConnectionProvider p = ConnectionProvider.fixed("test", 1);
-		AtomicReference<Channel> ch1 = new AtomicReference<>();
-		AtomicReference<Channel> ch2 = new AtomicReference<>();
-
-		HttpResponseStatus r =
-				HttpClient.create(p)
-				          .doOnResponse((res, c) -> ch1.set(c.channel()))
-				          .wiretap(true)
-				          .get()
-				          .uri("http://localhost:" + tomcat.port() + "/status/404")
-				          .responseSingle((res, buf) -> buf.thenReturn(res.status()))
-				          .block(Duration.ofSeconds(30));
-
-		HttpClient.create(p)
-		          .doOnResponse((res, c) -> ch2.set(c.channel()))
-		          .wiretap(true)
-		          .get()
-		          .uri("http://localhost:" + tomcat.port() + "/status/404")
-		          .responseSingle((res, buf) -> buf.thenReturn(res.status()))
-		          .block(Duration.ofSeconds(30));
-
-		AtomicBoolean same = new AtomicBoolean();
-
-		same.set(ch1.get() == ch2.get());
-
-		Assert.assertTrue(same.get());
-
-		Assert.assertEquals(r, HttpResponseStatus.NOT_FOUND);
-		p.dispose();
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void disableChunkImplicitDefault() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		ConnectionProvider p = ConnectionProvider.fixed("test", 1);
-		HttpClient client =
-				HttpClient.create(p)
-				          .tcpConfiguration(tcpClient -> tcpClient.host("localhost"))
-				          .port(tomcat.port())
-				          .wiretap(true);
-
-		Tuple2<HttpResponseStatus, Channel> r =
-				client.get()
-				      .uri("/status/404")
-				      .responseConnection((res, conn) -> Mono.just(res.status())
-				                                             .delayUntil(s -> conn.inbound().receive())
-				                                             .zipWith(Mono.just(conn.channel())))
-				      .blockLast(Duration.ofSeconds(30));
-
-		assertThat(r).isNotNull();
-
-		Channel r2 =
-				client.get()
-				      .uri("/status/404")
-				      .responseConnection((res, conn) -> Mono.just(conn.channel())
-				                                             .delayUntil(s -> conn.inbound().receive()))
-				      .blockLast(Duration.ofSeconds(30));
-
-		assertThat(r2).isNotNull();
-
-		Assert.assertSame(r.getT2(), r2);
-
-		Assert.assertEquals(r.getT1(), HttpResponseStatus.NOT_FOUND);
-		p.dispose();
-
-		tomcat.stop();
-	}
-
-	@Test
-	public void contentHeader() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		ConnectionProvider fixed = ConnectionProvider.fixed("test", 1);
-		HttpClient client =
-				HttpClient.create(fixed)
-				          .wiretap(true)
-				          .headers(h -> h.add("content-length", "1"));
-
-		HttpResponseStatus r =
-				client.request(HttpMethod.GET)
-				      .uri("http://localhost:" + tomcat.port())
-				      .send(ByteBufFlux.fromString(Mono.just(" ")))
-				      .responseSingle((res, buf) -> Mono.just(res.status()))
-				      .block(Duration.ofSeconds(30));
-
-		client.request(HttpMethod.GET)
-		      .uri("http://localhost:" + tomcat.port())
-		      .send(ByteBufFlux.fromString(Mono.just(" ")))
-		      .responseSingle((res, buf) -> Mono.just(res.status()))
-		      .block(Duration.ofSeconds(30));
-
-		Assert.assertEquals(r, HttpResponseStatus.BAD_REQUEST);
-		fixed.dispose();
-
-		tomcat.stop();
 	}
 
 	@Test
