+++ src/test/java/reactor/netty/tcp/TcpClientTests.java
@@ -54,10 +54,8 @@ import reactor.netty.Connection;
 import reactor.netty.DisposableServer;
 import reactor.netty.NettyOutbound;
 import reactor.netty.SocketUtils;
-import reactor.netty.TomcatServer;
 import reactor.netty.channel.AbortedException;
 import reactor.netty.channel.ChannelOperations;
-import reactor.netty.http.client.HttpClient;
 import reactor.netty.resources.ConnectionProvider;
 import reactor.netty.resources.LoopResources;
 import reactor.test.StepVerifier;
@@ -555,32 +553,6 @@ public class TcpClientTests {
 
 		assertThat(duration, is(greaterThanOrEqualTo(500L)));
 		client.disposeNow();
-	}
-
-	@Test
-	public void nettyNetChannelAcceptsNettyChannelHandlers() throws Exception {
-		TomcatServer tomcat = new TomcatServer();
-		tomcat.createDefaultContext();
-		tomcat.start();
-
-		HttpClient client = HttpClient.create()
-		                              .port(tomcat.port())
-		                              .wiretap(true);
-
-		final CountDownLatch latch = new CountDownLatch(1);
-		String response = client.get()
-		                        .uri("/?q=test%20d%20dq")
-		                        .responseContent()
-		                        .aggregate()
-		                        .asString()
-		                        .doOnSuccess(v -> latch.countDown())
-		                        .block(Duration.ofSeconds(30));
-
-		assertTrue("Latch didn't time out", latch.await(15, TimeUnit.SECONDS));
-		assertNotNull(response);
-		assertTrue(response.contains("q=test%20d%20dq"));
-
-		tomcat.stop();
 	}
 
 	@Test
